- Changed member variable keeping track of total samples to 64 bits long so doesn't overflow - Added normalization when loading training data for positions and tMax values - Added gather step kernel function. Has not been integrated to `train_step` yet.

+++ src/pbrt/wavefront/integrator.cpp
@@ -354,7 +354,7 @@ namespace pbrt
                     size_t samplesPos = line.find("samples=");
                     if (samplesPos != std::string::npos) {
                         try {
-                            rayLogSampleCnt = std::stoi(line.substr(samplesPos + 8));
+                            rayLogSampleCnt = std::stoll(line.substr(samplesPos + 8));
                         } catch (...) {}
                     }
                 }
